requirements.txt
fastapi
uvicorn
gtts
langdetect
datetime
google-cloud-texttospeech
google-au
protobuf==3.13.*
